refactor: reduce int type change

+++ controllers/casbin_adapter.go
@@ -48,7 +48,7 @@ func (c *ApiController) GetCasbinAdapters() {
 			return
 		}
 
-		paginator := pagination.SetPaginator(c.Ctx, limit, int64(count))
+		paginator := pagination.SetPaginator(c.Ctx, limit, count)
 		adapters, err := object.GetPaginationCasbinAdapters(owner, organization, paginator.Offset(), limit, field, value, sortField, sortOrder)
 		if err != nil {
 			c.ResponseError(err.Error())
